+++ resources/js/survey/views/layout/Page.vue
@@ -16,8 +16,9 @@
 
     <div class="bg-warning w-100 d-flex align-items-center mt-auto" style="height:60px">
       <b-container>
+
         <div class="text-md-right text-center ">
-          <b-img fluid src="/img/icon-aecom-kpmg.svg"></b-img>
+          <b-img style="height:40px" fluid src="/img/icon-aecom-kpmg.svg"></b-img>
 
         </div>
       </b-container>