+++ resources/js/survey/views/survey/SocialData.vue
@@ -2,66 +2,47 @@
   <div class="">
     <div v-if="!isLoading" class=" d-flex justify-content-center py-3">
 
-      <card-survey v-if="step == 1" action="Lanjut" :options="options.gender" @onNext="nextStep(2,5)" @onChange="input.gender = $event">
+      <card-survey v-if="step == 1" :step="step" action="Lanjut" :options="options.gender" @onNext="nextStep(2)" @onChange="input.gender = $event">
         <template v-slot:question>
           Jenis kelamin
         </template>
       </card-survey>
-      <card-survey v-if="step == 2" action="Lanjut" :options="options.age" @onNext="nextStep(3,10)" @onChange="input.age = $event">
+      <card-survey v-if="step == 2" action="Lanjut" :options="options.age" @onBack="backStep(1)" @onNext="nextStep(3)" @onChange="input.age = $event">
         <template v-slot:question>
           Umur
         </template>
       </card-survey>
-      <card-survey v-if="step == 3" action="Lanjut" :options="options.job" @onNext="nextStep(4,15)" @onChange="input.job = $event">
+      <card-survey v-if="step == 3" action="Lanjut" :options="options.job" @onBack="backStep(2)" @onNext="nextStep(4)" @onChange="input.job = $event">
         <template v-slot:question>
           Pekerjaan
         </template>
       </card-survey>
-      <card-survey v-if="step == 4" action="Lanjut" :options="options.income" @onNext="nextStep(5,20)" @onChange="input.income = $event">
+      <card-survey v-if="step == 4" action="Lanjut" :options="options.income" @onBack="backStep(3)" @onNext="nextStep(5)" @onChange="input.income = $event">
         <template v-slot:question>
           Pendapatan per bulan
         </template>
       </card-survey>
-      <b-col v-if="step == 5" lg="4" md="6" sm="8">
-        <b-card class="shadow-sm mb-3">
-          <div>
-            <p>
-              Berapa alokasi rata-rata biaya transportasi anda per bulan?
-            </p>
-            <vue-slider v-model="input.avg_transportation_cost" :height="6" :min="0" :max="100" :interval="1" :tooltip-formatter="sliderTooltipFormat">
-              <!-- <template v-slot:mark="{ pos, label }">
-              <div class="custom-mark" :style="{ left: `${pos}%` }">
-                {{ label }}
-              </div>
-            </template> -->
-              <!-- <template v-slot:step="{ active, value }">
-              <div :class="['vue-slider-mark-label', 'custom-label', { active }]">{{ value }}</div>
-            </template> -->
-              <!-- <template v-slot:label="{ active, value }">
-              <div :class="['vue-slider-mark-label', 'custom-label', { active }]">{{ value }}</div>
-            </template> -->
-            </vue-slider>
-            <div class="text-center">
-              <span v-if="income.selected[0].max == null">≤</span>
-              <span v-if="income.selected[0].min != null"> Rp. {{income.selected[0].min * input.avg_transportation_cost/100 | currency}}</span>
+      <card-survey-slider v-if="step == 5" action="Lanjut" :height="6" :min="0" :max="100" :interval="1" :sliderTooltipFormat="'{value}%'" @onChange="input.avg_transportation_cost = $event" @onBack="backStep(4)" @onNext="nextStep(6)">
+        <template v-slot:question>
+          Berapa alokasi rata-rata biaya transportasi anda per bulan?
+        </template>
+        <template v-slot:info>
+          <span v-if="income.selected[0].max == null">≤</span>
+          <span v-if="income.selected[0].min != null"> Rp. {{income.selected[0].min * input.avg_transportation_cost/100 | currency}}</span>
 
-              <span v-if="income.selected[0].min != null && income.selected[0].max != null"> - </span>
-              <span v-if="income.selected[0].min == null">></span>
+          <span v-if="income.selected[0].min != null && income.selected[0].max != null"> - </span>
+          <span v-if="income.selected[0].min == null">></span>
 
-              <span v-if="income.selected[0].max != null"> Rp. {{income.selected[0].max * input.avg_transportation_cost/100 | currency}}</span>
-            </div>
-          </div>
-        </b-card>
-        <b-btn v-if="input.avg_transportation_cost != null" variant="primary" class="font-weight-bold" block @click="nextStep(6,25)">Lanjut</b-btn>
-      </b-col>
-      <card-survey v-if="step == 6" action="Lanjut" :options="options.vehicle" @onNext="nextStep(7,30)" @onChange="input.vehicle = $event">
+          <span v-if="income.selected[0].max != null"> Rp. {{income.selected[0].max * input.avg_transportation_cost/100 | currency}}</span> </template>
+      </card-survey-slider>
+      <card-survey v-if="step == 6" action="Lanjut" :options="options.vehicle" @onNext="nextStep(7)" @onBack="backStep(5)" @onChange="input.vehicle = $event">
         <template v-slot:question>
           Apa kendaraan yang punya?
         </template>
       </card-survey>
-      <card-survey v-if="step == 7" action="Lanjut" :options="options.is_singkawang_domicile" @onNext="nextStep(8,35)" @onChange="input.is_singkawang_domicile = $event">
+      <card-survey v-if="step == 7" action="Lanjut" :options="options.is_singkawang_domicile" @onNext="nextStep(8)" @onBack="backStep(6)" @onChange="input.is_singkawang_domicile = $event">
         <template v-slot:question>
-          Apakah anda berdomisili di Singkawang dan sekitarnya?
+          Apakah anda berdomisili di Singkawang dan sekitarnya (Kabupaten Bengkayang, Sambas dan Mempawah)?
         </template>
       </card-survey>
       <b-col v-if="step == 8" lg="4" md="6" sm="8">
@@ -86,7 +67,12 @@
             </b-form-group>
           </div>
         </b-card>
-        <b-btn v-if="input.domicile_city != '' && input.domicile_province != ''" variant="primary" class="font-weight-bold" block @click="submit">Kirim</b-btn>
+        <div>
+          <b-btn variant="outline-secondary" @click="backStep(7)">
+            <i class="ri-arrow-left-circle-line align-middle"></i> Kembali
+          </b-btn>
+          <b-btn v-if="input.domicile_city != '' && input.domicile_province != ''" variant="primary" class="font-weight-bold float-right px-4" @click="submit">Kirim <i class="ri-send-plane-line align-middle"></i></b-btn>
+        </div>
       </b-col>
 
     </div>
@@ -118,6 +104,7 @@
 </template>
 <script>
   import CardSurvey from '../../components/CardSurvey.vue'
+  import CardSurveySlider from '../../components/CardSurveySlider.vue'
   import animationData from "../../asset/light_bulb.js";
   import {
     AuthRespondent
@@ -129,7 +116,8 @@
   export default {
     name: 'SocialData',
     components: {
-      CardSurvey
+      CardSurvey,
+      CardSurveySlider
     },
     mixins: [AuthRespondent, MethodsSurvey],
     props: ["routerData"],
@@ -141,6 +129,7 @@
           animationData: animationData,
           loop: true
         },
+        stepProgress: [5, 10, 15, 20, 25, 30, 35],
         isLoading: false,
         submitting: false,
         animationSpeed: 1,
@@ -185,6 +174,7 @@
     created() {
       this.getData();
       this.$emit("childinit", this.routerData);
+      this.stepProgress.unshift(this.routerData.progress)
     },
     watch: {
       'input.income': function(newVal, oldVal) {